Implemented "Gravity" setting toggle

+++ tetresse-legacy/scripts/game.js
@@ -527,31 +527,29 @@ class Game {
             v.classList.add(this.name + "-menu-break");
             v.innerHTML = "Training";
             // Gravity setting
-            if (true) {
-                var g, e;
-                g = addChild(training, training.id + "-container-0", "div");
-                g.classList.add(this.name + "-menu-keybind-group");
-                e = addChild(g, g.id + "-element-grav", "div");
-                e.classList.add(this.name + "-menu-keybinds-element");
-                e.classList.add(this.name + "-al");
-                v = addChild(e, e.id + "-question", "div");
-                v.classList.add(this.name + "-question");
-                v.classList.add(this.name + "-al");
-                v.classList.add(this.name + "-menu-keybinds-button");
-                v.title = "Turns gravity on or off";
-                v = addChild(e, e.id + "-text", "div");
-                v.classList.add(this.name + "-al");
-                v.innerHTML = "Gravity";
-                v = addChild(e, e.id + "-checkbox", "input");
-                v.type = "checkbox";
-                v.classList.add(this.name + "-ar", this.name + "-menu-keybinds-checkbox");
-                v.checked = this.settings.gravityEnabled;
-                addEvent(v, "change", function(e) {
-                    var board = games[Game.getGameNumber(e.target.id)];
-                    board.settings.gravityEnabled = e.target.checked;
-                    setCookie("settings", JSON.stringify(board.settings), 1000);
-                });
-            }
+            var g, e;
+            g = addChild(training, training.id + "-container-0", "div");
+            g.classList.add(this.name + "-menu-keybind-group");
+            e = addChild(g, g.id + "-element-grav", "div");
+            e.classList.add(this.name + "-menu-keybinds-element");
+            e.classList.add(this.name + "-al");
+            v = addChild(e, e.id + "-question", "div");
+            v.classList.add(this.name + "-question");
+            v.classList.add(this.name + "-al");
+            v.classList.add(this.name + "-menu-keybinds-button");
+            v.title = "Turns gravity on or off";
+            v = addChild(e, e.id + "-text", "div");
+            v.classList.add(this.name + "-al");
+            v.innerHTML = "Gravity";
+            v = addChild(e, e.id + "-checkbox", "input");
+            v.type = "checkbox";
+            v.classList.add(this.name + "-ar", this.name + "-menu-keybinds-checkbox");
+            v.checked = this.settings.gravityEnabled;
+            addEvent(v, "change", function(e) {
+                var board = games[Game.getGameNumber(e.target.id)];
+                board.settings.gravityEnabled = e.target.checked;
+                setCookie("settings", JSON.stringify(board.settings), 1000);
+            });
 
             // menu stuffs
             if (true) {
@@ -2068,9 +2066,11 @@ class GravityTimer {
                 board.gravNum--;
             if (board.gravNum != 0)
                 return;
-            
-            board.piece.drop();
-            
+           
+            if (board.settings.gravityEnabled) {
+                board.piece.drop();
+            }
+
             if (board.gravTimer != null)
                 board.gravTimer = new GravityTimer(board);
         }
